requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
dnspython==2.3.0
Flask==2.2.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.2
PyJWT==2.6.0
pymongo==4.3.3
pytz==2022.7
requests==2.28.1
twilio==7.16.0
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
zipp==3.12.0
